requirements.txt
azure_storage==0.37.0
openai==0.27.8
pandas==1.5.1
python-dotenv==1.1.1
